refactor: move security module nearer to the root, away form helpers
