+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Security/Features/SinunaSecurityFeature.cs
@@ -4,9 +4,10 @@ using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.IdentityModel.Tokens;
 using NetDevPack.Security.JwtExtensions;
+using VirtualFinland.UserAPI.Helpers;
 using JwksExtension = VirtualFinland.UserAPI.Helpers.Extensions.JwksExtension;
 
-namespace VirtualFinland.UserAPI.Helpers.Security.Features;
+namespace VirtualFinland.UserAPI.Security.Features;
 
 public class SinunaSecurityFeature : ISecurityFeature
 {
