+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Security/Features/SuomiFiSecurityFeature.cs
@@ -3,9 +3,10 @@ using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.IdentityModel.Tokens;
 using NetDevPack.Security.JwtExtensions;
+using VirtualFinland.UserAPI.Helpers;
 using JwksExtension = VirtualFinland.UserAPI.Helpers.Extensions.JwksExtension;
 
-namespace VirtualFinland.UserAPI.Helpers.Security.Features;
+namespace VirtualFinland.UserAPI.Security.Features;
 
 public class SuomiFiSecurityFeature : ISecurityFeature
 {
