+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Security/Features/TestBedConsentProviderConfig.cs
@@ -2,7 +2,7 @@ using System.Text.Json;
 using System.Text.Json.Serialization;
 using Microsoft.IdentityModel.Tokens;
 
-namespace VirtualFinland.UserAPI.Helpers.Security.Features;
+namespace VirtualFinland.UserAPI.Security.Features;
 
 public class TestBedConsentProviderConfig : IConsentProviderConfig
 {
