+++ VirtualFinland.UserAPI/src/VirtualFinland.UsersAPI/Security/Models/JwtTokenResult.cs
@@ -1,4 +1,4 @@
-namespace VirtualFinland.UserAPI.Helpers.Security;
+namespace VirtualFinland.UserAPI.Security.Models;
 
 public class JwtTokenResult
 {
