+++ VirtualFinland.UsersAPI.UnitTests/Tests/Security/AuthenticationTests.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Logging;
 using Moq;
 using Moq.Protected;
 using VirtualFinland.UserAPI.Exceptions;
-using VirtualFinland.UserAPI.Helpers.Security;
+using VirtualFinland.UserAPI.Security;
 using VirtualFinland.UserAPI.Helpers.Services;
 using VirtualFinland.UsersAPI.UnitTests.Helpers;
